Filter out 'resolved' field from shrink-wrap files.
The field is *meant* to ensure packages are installed from the correct
registry, but when building on shared equipment (i.e. build servers)
the resolved field causes a range of problems.

This patch filters out the field so installs will come from the user's
selected registry or proxy and not bypass it.